fix: rename middleware to proxy to match nextjs convention
